release: changed the old BlocklyComponent class component with a new functional component
