+++ examples/blockly-react/src/Blockly/BlocklyComponent.jsx
@@ -21,55 +21,54 @@
  * @author samelh@google.com (Sam El-Husseini)
  */
 
-import React from 'react';
-import './BlocklyComponent.css';
+ import React from 'react';
+ import './BlocklyComponent.css';
+ import {useEffect, useRef} from 'react';
 
-import Blockly from 'blockly/core';
-import locale from 'blockly/msg/en';
-import 'blockly/blocks';
+ import Blockly from 'blockly/core';
+ import BlocklyJS from 'blockly/javascript';
+ import locale from 'blockly/msg/en';
+ import 'blockly/blocks';
+ 
+ Blockly.setLocale(locale);
+ 
+ function BlocklyComponent(props) {
+    const blocklyDiv = useRef();
+    const toolbox = useRef();
+    let primaryWorkspace = useRef();
 
-Blockly.setLocale(locale);
-
-class BlocklyComponent extends React.Component {
-    constructor(props) {
-        super(props);
-        this.blocklyDiv = React.createRef();
-        this.toolbox = React.createRef();
-    }
-
-    componentDidMount() {
-        const { initialXml, children, ...rest } = this.props;
-        this.primaryWorkspace = Blockly.inject(
-            this.blocklyDiv.current,
-            {
-                toolbox: this.toolbox.current,
-                ...rest
-            },
+    const generateCode = () => {
+        console.log(primaryWorkspace.current.workspace);
+        var code = BlocklyJS.workspaceToCode(
+          primaryWorkspace.current
         );
-
-        if (initialXml) {
-            Blockly.Xml.domToWorkspace(Blockly.Xml.textToDom(initialXml), this.primaryWorkspace);
-        }
+        console.log(code);
     }
+ 
+    useEffect(() => {
+        const { initialXml, children, ...rest } = props;
+            primaryWorkspace.current = Blockly.inject(
+                blocklyDiv.current,
+                {
+                    toolbox: toolbox.current,
+                    ...rest
+                },
+            );
 
-    get workspace() {
-        return this.primaryWorkspace;
-    }
-
-    setXml(xml) {
-        Blockly.Xml.domToWorkspace(Blockly.Xml.textToDom(xml), this.primaryWorkspace);
-    }
-
-    render() {
-        const { children } = this.props;
-
-        return <React.Fragment>
-            <div ref={this.blocklyDiv} id="blocklyDiv" />
-            <xml xmlns="https://developers.google.com/blockly/xml" is="blockly" style={{ display: 'none' }} ref={this.toolbox}>
-                {children}
-            </xml>
-        </React.Fragment>;
-    }
+            if (initialXml) {
+                Blockly.Xml.domToWorkspace(Blockly.Xml.textToDom(initialXml), primaryWorkspace.current);
+            }
+    }, [primaryWorkspace, toolbox, blocklyDiv, props]);
+ 
+    return (
+    <React.Fragment>
+        <button onClick={generateCode}>Convert</button>
+        <div ref={blocklyDiv} id="blocklyDiv" />
+        <div style={{ display: 'none' }} ref={toolbox}>
+            {props.children}
+        </div>
+    </React.Fragment>);
 }
-
+ 
 export default BlocklyComponent;
+
